Forked and popped specs2 in
We use specs2, and obviously we don't use com.twitter as our prefix. So,
this makes the fork and does all the changes we need to run it
ourselves.

Thanks to @twitter for making this available - it was invaluable in my
training while at Twitter and I just still want to use it.